% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{boostr}
\alias{boostr}
\alias{boostr-package}
\title{Boost (or bag) an estimation procedure with any reweighter or aggregator.}
\description{
boostr provides a modular framework that return the focus of ensemble learning
back to 'learning' (instead of programming).
}

