% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibody.R
\name{ab}
\alias{ab}
\title{Estimate antibody titres}
\usage{
ab()
}
\arguments{
\item{timesteps}{Number of timesteps to output titres for}

\item{dose_timesteps}{Timesteps of doses}

\item{init_titres}{Titres at initital does or subsequent boost times}

\item{prop_short}{Proportion of the response that is short lived for each dose period}

\item{dur_short}{Duration of the short-lived response for each dose period}

\item{dur_long}{Duration of the long-lived response for each dose period}

\item{cpp}{Use cpp titre model}
}
\value{
Vector of antibody tires for 1:timesteps
}
\description{
Estimate antibody titres
}
