% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostr-package.R
\docType{package}
\name{boostr-package}
\alias{boostr}
\alias{boostr-package}
\title{boostr: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mrc-ide.github.io/boostr/}
}

}
\keyword{internal}
